requirements.txt
python-dotenv==1.0.1
redis==5.2.1
